% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_subgrp_data.R
\name{generate_subgrp_data}
\alias{generate_subgrp_data}
\title{Generate Subgroup Data-sets}
\usage{
generate_subgrp_data(n = 800, seed = 513413, family, ...)
}
\arguments{
\item{n}{sample size (default=800)}

\item{seed}{seed number (default=513413)}

\item{family}{Outcome type ("gaussian", "binomial", "survival")}

\item{...}{Any additional parameters, not currently passed through.}
}
\value{
Simulation data set (Y=outcome, A=treatment, X=covariates)
}
\description{
Simulation/real data-sets; useful for testing new models and PRISM configurations.
}
