% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsefulPredictors.R
\name{getUsefulPredictors}
\alias{getUsefulPredictors}
\title{Get Useful Predictors}
\usage{
getUsefulPredictors(x)
}
\arguments{
\item{x}{partykit model object}
}
\value{
Variable names that define the subgroups in the partykit model object
}
\description{
Extract the variables used in subgroup definitions (partykit objects)
}
