% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_rules.R
\name{list_rules}
\alias{list_rules}
\title{Get Useful Predictors}
\usage{
list_rules(x, i = NULL, ...)
}
\arguments{
\item{x}{partykit model object}

\item{i}{node (default=NULL)}

\item{...}{Other arguments}
}
\value{
Subgroup rules/definitions. Based on partykit:::.list.rules.party
}
\description{
Extract the variables used in subgroup definitions (partykit objects)
}
