% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_template.R
\name{model_template}
\alias{model_template}
\title{Model Template: Generates template functions for PRISM algorithm}
\usage{
model_template(type)
}
\arguments{
\item{type}{Type of model template, accepts "filter", "ple", "submod", and "param".}
}
\value{
Function template for chosen model type.
}
\description{
Helper function that generates template functions for the individual components of 
PRISM (filter, ple, submod, param). Useful for creating user-specific models, for 
example a different ple model for estimating counterfactuals.
}
\examples{
model_template(type="filter")
model_template(type="ple")
model_template(type="submod")
model_template(type="param")
}
