% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ple_bart.R
\name{predict.ple_bart}
\alias{predict.ple_bart}
\title{Predict Patient-level Estimates: BART}
\usage{
\method{predict}{ple_bart}(object, newdata, ...)
}
\arguments{
\item{object}{Trained BART model(s).}

\item{newdata}{Data-set to make predictions at.}

\item{...}{Any additional parameters, not currently passed through.}
}
\value{
Data-frame with predictions of (E(Y|X,1), E(Y|X,0), E(Y|X,1)-E(Y|X,0))
}
\description{
Get estimates of (E(Y|X,A=1), E(Y|X,A=0), E(Y|X,A=1)-E(Y|X,A=0)) using trained
BART model(s).
}
\examples{
library(StratifiedMedicine)

## Continuous ##
dat_ctns = generate_subgrp_data(family="gaussian")
Y = dat_ctns$Y
X = dat_ctns$X
A = dat_ctns$A
\donttest{
mod1 = ple_bart(Y, A, X, Xtest=X)
summary(mod1$mu_train)
summary(predict(mod1, newdata=data.frame(A,X)))
summary(predict(mod1, newdata=data.frame(X)))
}
}
