% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ple_ranger.R
\name{predict.ple_ranger}
\alias{predict.ple_ranger}
\title{Predict Patient-level Estimates: Ranger}
\usage{
\method{predict}{ple_ranger}(object, newdata, oob = TRUE, ...)
}
\arguments{
\item{object}{Trained random forest (ranger) model(s).}

\item{newdata}{Data-set to make predictions at. For training data, should
include covariates (X) and treatment (A) for oob predictions.}

\item{oob}{Use out-of-bag predictions (default=TRUE unless newdata does not contain
treatment variable A).}

\item{...}{Any additional parameters, not currently passed through.}
}
\value{
Data-frame with predictions of (E(Y|X,A=1), E(Y|X,A=0), E(Y|X,A=1)-E(Y|X,A=0))
}
\description{
Get estimates of (E(Y|X,A=1), E(Y|X,A=0), E(Y|X,A=1)-E(Y|X,A=0)) using trained random
forest (ranger) model(s).
}
\examples{
library(StratifiedMedicine)
## Continuous ##
dat_ctns = generate_subgrp_data(family="gaussian")
Y = dat_ctns$Y
X = dat_ctns$X
A = dat_ctns$A


# Default (treatment-specific ranger models) #
mod1 = ple_ranger(Y, A, X, Xtest=X)
summary( predict(mod1, newdata=data.frame(A,X) ) ) # oob predictions for training
summary( predict(mod1, newdata=data.frame(X) ) ) # new-predictions, no oob here


}
