% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submod_rpart.R
\name{predict.submod_rpart}
\alias{predict.submod_rpart}
\title{Predict submod: rpart}
\usage{
\method{predict}{submod_rpart}(object, newdata, ...)
}
\arguments{
\item{object}{Trained ctree model.}

\item{newdata}{Data-set to make predictions at.}

\item{...}{Any additional parameters, not currently passed through.}
}
\value{
Identified subgroups with subgroup-specific predictions of E(Y|X) or PLE(X).
\itemize{
  \item Subgrps - Identified subgroups
  \item pred - Predictions, E(Y|X) or PLE(X) by subgroup.
}
}
\description{
Predict subgroups and obtain subgroup-specific estimates, E(Y|X) or PLE(X), for a
trained ctree model (depends on if outcome_PLE argument)
}
\examples{

\donttest{
library(StratifiedMedicine)

## Continuous ##
dat_ctns = generate_subgrp_data(family="gaussian")
Y = dat_ctns$Y
X = dat_ctns$X
A = dat_ctns$A

res_rpart = submod_rpart(Y, A, X, Xtest=X)
# Predict subgroups / estimates #
out = predict(res_rpart, newdata=X)
}

}
