% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubMod_CART.R
\name{SubMod_CART}
\alias{SubMod_CART}
\title{Subgroup Identification: CART}
\usage{
SubMod_CART(Y, A, X, Xtest, mu_train, minbucket = floor(dim(X)[1] *
  0.05), maxdepth = 4, outcome_PLE = FALSE, family = "gaussian", ...)
}
\arguments{
\item{Y}{The outcome variable. Must be numeric or survival (ex; Surv(time,cens) )}

\item{A}{Treatment variable. (a=1,...A)}

\item{X}{Covariate matrix. Must be numeric.}

\item{Xtest}{Test set}

\item{mu_train}{Patient-level estimates (See PLE_models)}

\item{minbucket}{Minimum number of observations in a tree node.
Default = floor( dim(train)[1]*0.05  )}

\item{maxdepth}{Maximum depth of any node in the tree (default=4)}

\item{outcome_PLE}{If TRUE, use PLE as outcome (mu_train must contain PLEs).
Else use observed outcome Y}

\item{family}{Outcome type ("gaussian", "binomial"), default is "gaussian"}

\item{...}{Any additional parameters, not currently passed through.}
}
\value{
CART model, predictions, identified subgroups, and subgroup rules/definitions.
 \itemize{
  \item mod - CART model object
  \item Subgrps.train - Identified subgroups (training set)
  \item Subgrps.test - Identified subgroups (test set)
  \item pred.train - Predictions (training set)
  \item pred.test - Predictions (test set)
  \item Rules - Subgroups rules/definitions
}
}
\description{
Uses the CART algorithm to identify subgroups. Usable for continuous and binary outcomes.
Option to use the observed outcome or PLEs for subgroup identification.
}
\examples{
library(StratifiedMedicine)

## Continuous ##
dat_ctns = generate_subgrp_data(family="gaussian")
Y = dat_ctns$Y
X = dat_ctns$X
A = dat_ctns$A

res_CART1 = SubMod_CART(Y, A, X, Xtest=X, family="gaussian")
res_CART2 = SubMod_CART(Y, A, X, Xtest=X, maxdepth=2, minbucket=100, family="gaussian")
plot(res_CART1$mod)
plot(res_CART2$mod)

}
\seealso{
\code{\link{PRISM}}, \code{\link{rpart}}
}
