% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submod_prune.R
\name{submod_prune}
\alias{submod_prune}
\title{Subgroup Identification: Pruning}
\usage{
submod_prune(Y, A, X, Xtest, type, submod0, submod = NULL,
  hyper = NULL, mu_train = NULL, family = "gaussian", ...)
}
\arguments{
\item{Y}{The outcome variable. Must be numeric or survival (ex; Surv(time,cens) )}

\item{A}{Treatment variable. (a=1,...A)}

\item{X}{Covariate space.}

\item{Xtest}{Test set}

\item{type}{Type of pruning. Options include "OTR" and "submod2X".}

\item{submod0}{Initial subgroup model fit.}

\item{submod}{Subgroup identification (submod) function. Maps the observed data and/or
PLEs to subgroups.}

\item{hyper}{Hyper-parameters (must be list). Default is NULL.}

\item{mu_train}{Patient-level estimates (See PLE_models). Default=NULL}

\item{family}{Outcome type ("gaussian", "binomial", "survival"). Default="gaussian".}

\item{...}{Any additional parameters, not currently passed through.}
}
\value{
Trained subgroup model and subgroup predictions/estimates for train/test sets.

 \itemize{
  \item mod - trained subgroup model
  \item Subgrps.train - Identified subgroups (training set)
  \item Subgrps.test - Identified subgroups (test set)
  \item pred.train - Predictions (training set)
  \item pred.test - Predictions (test set)
  \item Rules - Definitions for subgroups, if provided in submod output.
}
}
\description{
Function to prune a trained subgroup model. Can be used within PRISM. This is currently
experimental and includes two approaches: (1) Pruning initial subgroups based on optimal
treatment regime loss function, and (2) "Double" Subgroup model, which uses the initial
discovered subgroups as the input covariate in the subgroup model.
}
