% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISM.R
\name{summary.PRISM}
\alias{summary.PRISM}
\title{PRISM: Patient Response Identifier for Stratified Medicine (Summary)}
\usage{
\method{summary}{PRISM}(object, round_est = 4, round_SE = 4, round_CI = 4, ...)
}
\arguments{
\item{object}{Trained PRISM model.}

\item{round_est}{Rounding for trt ests (default=4)}

\item{round_SE}{Rounding for trt SEs (default=4)}

\item{round_CI}{Rounding for trt CIs (default=4)}

\item{...}{Any additional parameters, not currently passed through.}
}
\value{
List of key PRISM outputs: (1) Configuration, (2) Variables that pass filter 
(if filter is used), (3) Number of Identified Subgroups, and (4) Parameter Estimates, 
SEs, and CIs for each subgroup/estimand
}
\description{
Summary for PRISM algorithm results. Outputs configuration, which variables pass the filter (if used),
subgroup summaries, and treatment effect estimates.
}
