% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ple_bart.R
\name{ple_bart}
\alias{ple_bart}
\title{Patient-level Estimates: BART}
\usage{
ple_bart(Y, A, X, Xtest, family = "gaussian", sparse = FALSE, K = 10,
  ...)
}
\arguments{
\item{Y}{The outcome variable. Must be numeric.}

\item{A}{Treatment variable. (a=1,...A)}

\item{X}{Covariate space.}

\item{Xtest}{Test set}

\item{family}{Outcome type ("gaussian", "binomial"), default is "gaussian"
which uses wbart (BART for continuous outcomes). Probit-based BART ("pbart") is
used for family="binomial"}

\item{sparse}{Whether to perform variable selection based on sparse Dirichlet prior
instead of uniform. See "gbart" in the BART R package for more details as well as
Linero 2016. Default is FALSE.}

\item{K}{For survival, coarse times per the quantiles 1/K,..,K/K. For AFT (abart),
K=10. For more accurate survival predictions, set K=100 (default in abart).}

\item{...}{Any additional parameters, not currently passed through.}
}
\value{
Trained BART model(s) and patient-level estimates
(E(Y|X,1), E(Y|X,0), E(Y|X,1)-E(Y|X,0)) for train/test sets.
 \itemize{
  \item mod - trained model(s)
  \item mu_train - Patient-level estimates (training set)
  \item mu_test - Patient-level estimates (test set)
}
}
\description{
Uses the BART algorithm (Chipman et al 2010; BART R package) to obtain patient-level
estimates. Used for continuous or binary outcomes. Covariate by treatment interactions
are automatically included in BART model (as in Hahn et al 2017).
}
\examples{
library(StratifiedMedicine)

## Continuous ##
dat_ctns = generate_subgrp_data(family="gaussian")
Y = dat_ctns$Y
X = dat_ctns$X
A = dat_ctns$A
train = data.frame(Y, A, X)

# BART #
\donttest{
require(BART)
mod1 = ple_bart(Y, A, X, Xtest=X)
summary(mod1$mu_train)
summary(predict(mod1, newdata=X))
}


}
