% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLE_causal_forest.R
\name{PLE_causal_forest}
\alias{PLE_causal_forest}
\title{Patient-level Estimates: Causal Forest}
\usage{
PLE_causal_forest(Y, A, X, Xtest, tune = FALSE, num.trees = 500,
  family = "gaussian", ...)
}
\arguments{
\item{Y}{The outcome variable. Must be numeric or survival (ex; Surv(time,cens) )}

\item{A}{Treatment variable. (a=1,...A)}

\item{X}{Covariate matrix. Must be numeric.}

\item{Xtest}{Test set}

\item{tune}{If TRUE, use grf automatic hyper-parameter tuning. If FALSE (default), no tuning.}

\item{num.trees}{Number of trees (default=500)}

\item{family}{Outcome type ("gaussian", "binomial"), default is "gaussian"}

\item{...}{Any additional parameters, not currently passed through.}
}
\value{
Patient-level estimates (E(Y|X,1), E(Y|X,0), E(Y|X,1)-E(Y|X,0)) for train/test sets.
 \itemize{
  \item mods - trained model(s)
  \item mu_train - Patient-level estimates (training set)
  \item mu_test - Patient-level estimates (test set)
}
}
\description{
Uses the causal forest algorithm (grf R package) to obtain patient-level estimates.
Used for continuous or binary outcomes, with output estimates of
E(Y|X,A=a) and E(Y|X,A=1)-E(Y|X,A=0) (PLE).
}
\examples{
library(StratifiedMedicine)

## Continuous ##
dat_ctns = generate_subgrp_data(family="gaussian")
Y = dat_ctns$Y
X = dat_ctns$X
A = dat_ctns$A
train = data.frame(Y, A, X)
# Outcome/treatment must be labeled as Y/A #

\donttest{
# Default #
mod1 = PLE_causal_forest(train=train, Xtest=X)
summary(mod1$mu_train$PLE)

# Tune hyper-parameters (uncomment to run) #
mod2 = PLE_causal_forest(train=train, Xtest=X, tune=TRUE)
summary(mod2$mu_train$PLE)

}


}
\seealso{
\code{\link{PRISM}}, \code{\link{causal_forest}}
}
