% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PRISM}
\alias{plot.PRISM}
\title{plot.PRISM}
\usage{
\method{plot}{PRISM}(x, type = "CI", ...)
}
\arguments{
\item{x}{PRISM object}

\item{type}{Type of plot (default="CI"). Alternatively, if the PRISM object used resampling,
type="resample" plots the resampling distributions.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Plot (ggplot2) object
}
\description{
Plots PRISM results, either forest plot (estimate with CIs) or resampling distribution.
}
\seealso{
\code{\link{PRISM}}
}
