% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRISM.R
\name{summary.PRISM}
\alias{summary.PRISM}
\title{PRISM: Patient Response Identifier for Stratified Medicine (Summary)}
\usage{
\method{summary}{PRISM}(object, ...)
}
\arguments{
\item{object}{Trained PRISM model.}

\item{...}{Any additional parameters, not currently passed through.}
}
\value{
List of key PRISM outputs: (1) Configuration, (2) Variables that pass filter 
(if filter is used), (3) Number of Identified Subgroups, and (4) Parameter Estimates, 
SEs, and CIs for each subgroup/estimand
}
\description{
Predictions for PRISM algorithm. Given the training set (Y,A,X) or new test set (Xtest),
output ple predictions and identified subgroups with correspond parameter estimates.
}
